Add stat_bt_reader.go and move some functions to stat_util.go
