+++ server/stats.go
@@ -29,42 +29,6 @@ import (
 // Limit the concurrent channels when processing in-memory cache data.
 const maxChannelSize = 50
 
-// ObsProp represents properties for a StatObservation.
-type ObsProp struct {
-	Mmethod string
-	Operiod string
-	Unit    string
-	Sfactor string
-}
-
-func tokenFn(
-	keyTokens map[string]*placeStatVar) func(rowKey string) (string, error) {
-	return func(rowKey string) (string, error) {
-		return keyTokens[rowKey].place + "^" + keyTokens[rowKey].statVar, nil
-	}
-}
-
-// Filter a list of source series given the observation properties.
-func filterSeries(in []*SourceSeries, prop *ObsProp) []*SourceSeries {
-	result := []*SourceSeries{}
-	for _, series := range in {
-		if prop.Mmethod != "" && prop.Mmethod != series.MeasurementMethod {
-			continue
-		}
-		if prop.Operiod != "" && prop.Operiod != series.ObservationPeriod {
-			continue
-		}
-		if prop.Unit != "" && prop.Unit != series.Unit {
-			continue
-		}
-		if prop.Sfactor != "" && prop.Sfactor != series.ScalingFactor {
-			continue
-		}
-		result = append(result, series)
-	}
-	return result
-}
-
 // GetStatValue implements API for Mixer.GetStatValue.
 func (s *Server) GetStatValue(ctx context.Context, in *pb.GetStatValueRequest) (
 	*pb.GetStatValueResponse, error) {
@@ -579,194 +543,4 @@ func (s *Server) GetStat(ctx context.Context, in *pb.GetStatRequest) (
 	}
 
 	return &pb.GetStatResponse{Stat: result}, nil
-}
-
-// triplesToStatsVar converts a Triples cache into a StatisticalVarible object.
-func triplesToStatsVar(
-	statsVarDcid string, triples *TriplesCache) (*StatisticalVariable, error) {
-	// Get constraint properties.
-	propValMap := map[string]string{}
-	for _, t := range triples.Triples {
-		if t.Predicate == "constraintProperties" {
-			propValMap[t.ObjectID] = ""
-		}
-	}
-	statsVar := StatisticalVariable{}
-	// Populate the field.
-	for _, t := range triples.Triples {
-		if t.SubjectID != statsVarDcid {
-			continue
-		}
-		object := t.ObjectID
-		switch t.Predicate {
-		case "typeOf":
-			if object != "StatisticalVariable" {
-				return nil, status.Errorf(
-					codes.Internal, "%s is not a StatisticalVariable", t.SubjectID)
-			}
-		case "statType":
-			statsVar.StatType = strings.Replace(object, "Value", "", 1)
-		case "populationType":
-			statsVar.PopType = object
-		case "measurementMethod":
-			statsVar.MeasurementMethod = object
-		case "measuredProperty":
-			statsVar.MeasuredProp = object
-		case "measurementDenominator":
-			statsVar.MeasurementDenominator = object
-		case "measurementQualifier":
-			statsVar.MeasurementQualifier = object
-		case "scalingFactor":
-			statsVar.ScalingFactor = object
-		case "unit":
-			statsVar.Unit = object
-		default:
-			if _, ok := propValMap[t.Predicate]; ok {
-				if statsVar.PVs == nil {
-					statsVar.PVs = map[string]string{}
-				}
-				statsVar.PVs[t.Predicate] = object
-			}
-		}
-	}
-	return &statsVar, nil
-}
-
-// getValue get the stat value from ObsTimeSeries.
-// When date is given, it get the value from the highest ranked source series
-// that has the date.
-// When date is not given, it get the latest value from the highest ranked
-// source series.
-func getValue(in *ObsTimeSeries, date string) (float64, error) {
-	if in == nil {
-		return 0, status.Error(codes.Internal, "Nil obs time series for getValue()")
-	}
-	sourceSeries := in.SourceSeries
-	sort.Sort(byRank(sourceSeries))
-	if date != "" {
-		for _, series := range sourceSeries {
-			if value, ok := series.Val[date]; ok {
-				return value, nil
-			}
-		}
-		return 0, status.Errorf(codes.NotFound, "No data found for date %s", date)
-	}
-	latestDate := ""
-	var result float64
-	for _, series := range sourceSeries {
-		for date, value := range series.Val {
-			if date > latestDate {
-				latestDate = date
-				result = value
-			}
-		}
-	}
-	if latestDate == "" {
-		return 0, status.Errorf(codes.NotFound,
-			"No stat data found for %s", in.PlaceDcid)
-	}
-	return result, nil
-}
-
-func (in *ObsTimeSeries) filterAndRank(prop *ObsProp) {
-	if in == nil {
-		return
-	}
-	series := filterSeries(in.SourceSeries, prop)
-	sort.Sort(byRank(series))
-	if len(series) > 0 {
-		in.Data = series[0].Val
-		in.ProvenanceDomain = series[0].ProvenanceDomain
-		in.ProvenanceURL = series[0].ProvenanceURL
-	}
-	in.SourceSeries = nil
-}
-
-// readStats reads and process BigTable rows in parallel.
-// Consider consolidate this function and bigTableReadRowsParallel.
-func readStats(
-	ctx context.Context,
-	btTable *bigtable.Table,
-	rowList bigtable.RowList,
-	keyTokens map[string]*placeStatVar) (
-	map[string]map[string]*ObsTimeSeries, error) {
-
-	dataMap, err := bigTableReadRowsParallel(
-		ctx, btTable, rowList, convertToObsSeries, tokenFn(keyTokens),
-	)
-	if err != nil {
-		return nil, err
-	}
-	result := map[string]map[string]*ObsTimeSeries{}
-	for token, data := range dataMap {
-		parts := strings.Split(token, "^")
-		place := parts[0]
-		statVar := parts[1]
-		if result[place] == nil {
-			result[place] = map[string]*ObsTimeSeries{}
-		}
-		if data == nil {
-			result[place][statVar] = nil
-		} else {
-			result[place][statVar] = data.(*ObsTimeSeries)
-		}
-	}
-	return result, nil
-}
-
-// readStats reads and process BigTable rows in parallel.
-// Consider consolidate this function and bigTableReadRowsParallel.
-func readStatsPb(
-	ctx context.Context,
-	btTable *bigtable.Table,
-	rowList bigtable.RowList,
-	keyTokens map[string]*placeStatVar) (
-	map[string]map[string]*pb.ObsTimeSeries, error) {
-
-	dataMap, err := bigTableReadRowsParallel(
-		ctx, btTable, rowList, convertToObsSeriesPb, tokenFn(keyTokens),
-	)
-	if err != nil {
-		return nil, err
-	}
-	result := map[string]map[string]*pb.ObsTimeSeries{}
-	for token, data := range dataMap {
-		parts := strings.Split(token, "^")
-		place := parts[0]
-		statVar := parts[1]
-		if result[place] == nil {
-			result[place] = map[string]*pb.ObsTimeSeries{}
-		}
-		if data == nil {
-			result[place][statVar] = nil
-		} else {
-			result[place][statVar] = data.(*pb.ObsTimeSeries)
-		}
-	}
-	return result, nil
-}
-
-// readStatCollection reads and process ObsCollection cache from BigTable
-// in parallel.
-func readStatCollection(
-	ctx context.Context,
-	btTable *bigtable.Table,
-	rowList bigtable.RowList,
-	keyTokens map[string]string) (
-	map[string]*pb.ObsCollection, error) {
-
-	dataMap, err := bigTableReadRowsParallel(
-		ctx, btTable, rowList, convertToObsCollection,
-		func(rowKey string) (string, error) {
-			return keyTokens[rowKey], nil
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-	result := map[string]*pb.ObsCollection{}
-	for token, data := range dataMap {
-		result[token] = data.(*pb.ObsCollection)
-	}
-	return result, nil
 }
